Fix repeating items in top projects leaderboard on landing page.
Fixes https://geyserteam.atlassian.net/browse/GT-972.
