+++ src/hooks/index.ts
@@ -1,4 +1,3 @@
-export * from './useProjects';
 export * from './useFundingFormState';
 export * from './useAllGeyserProjectEntries';
 export * from './useFundingFlow';
